requirements.txt
automata-lib==7.1.0
boto==2.49.0
colour==0.1.5
contourpy==1.0.7
cycler==0.11.0
dill==0.3.6
dtreeviz==1.4.0
fonttools==4.39.4
frozendict==2.3.8
graphviz==0.20.1
greenlet==2.0.2
iniconfig==2.0.0
joblib==1.2.0
kiwisolver==1.4.4
mariadb==1.1.6
matplotlib==3.7.1
multiprocess==0.70.14
networkx==3.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pathos==0.3.0
Pillow==9.5.0
pluggy==1.0.0
pox==0.3.2
ppft==1.7.6.6
pydot==1.4.2
pyparsing==3.0.9
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
SQLAlchemy==2.0.13
tabulate==0.9.0
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
z3==0.2.0
z3-solver==4.12.2.0
